requirements.txt
drjit==0.4.6
huggingface_hub==0.29.3
imageio==2.34.2
ipdb==0.13.13
lovely_tensors==0.1.16
matplotlib==3.7.1
mitsuba==3.5.2
numpy==2.2.4
open3d==0.18.0
opencv_python==4.7.0.72
Pillow==11.1.0
pytorch_lightning==1.9.0
PyYAML==6.0.2
scikit_image==0.19.3
scipy==1.15.2
torch==2.4.1
torchmetrics==1.4.0.post0
torchvision==0.19.1
tqdm==4.67.1
tqdm==4.66.4
wandb==0.19.9
xformers==0.0.29.post3
